Enable inviting more folks to a private conversation
More clear use and inndication of which keys are hashes and witch Base 64 strings.
Enable passing either kind of key to the Avatar component, and fix issues and display of identicons.

When adding a new contact go back to the previous page that got you there.

Tweak display of conversation header images