More tweaks to conversation joining flow
Show empty group icon in inbox while waiting for icon to sync.

Remove the send invitations block in a conversation once someone has joined.

Ass spinner to create button when creating a conversation from the create page.

Correct title when joining a private conversation, before it has synced.

When going to an existing converstation with a contact, or set of contacts, from the create page, prioritize exosting private conversations with them before jumping to an existing public conversation with that set of contacts.