Show appropriate title when someone joins a private conversation
For 1:1 conversation it should be the name of the conversation creator.
For group conversations it should be name of creator + # of people invited.

TODO: right now we only show the # of people invited until the conversation syncs, then it shows the profiles of the actual joined members. This could be confusing at the beginning when its only you and the creator. Would be ncie to store the # of people invited in the conversation cell from the beginning.

This is a breaking change because we removed name from conversation Properties